src/db_scripts/output_sdk/spaces/pyproject.toml
[tool.poetry]
name = "cisco-spaces-location-cloud-api-client"
version = "2.0.0"
description = "A client library for accessing Cisco Spaces Location Cloud API"
authors = []
readme = "README.md"
packages = [
    {include = "cisco_spaces_location_cloud_api_client"},
]
include = ["CHANGELOG.md", "cisco_spaces_location_cloud_api_client/py.typed"]


[tool.poetry.dependencies]
python = "^3.9"
httpx = ">=0.20.0,<0.29.0"
attrs = ">=22.2.0"
python-dateutil = "^2.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["F", "I", "UP"]
